\name{BayesPGLS-package}
\alias{BayesPGLS-package}
\alias{BayesPGLS}

\docType{package}

\title{
 Bayesian phylogenetic generalized least squares (PGLS).
}

\description{
 Provides functions to run Bayesian PGLS regressions including diagnostics.
}

\details{
\tabular{ll}{
	Package: \tab BayesPGLS \cr
	Type: \tab Package\cr
	Version: \tab 1.0.0\cr
	Date: \tab 2022-04-01\cr
	License: \tab GNU General Public Licence\cr
	LazyLoad: \tab yes\cr
  }
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\references{
FILL UP
}