\name{plot.BayesPGLS}
\alias{plot.BayesPGLS}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{plot.BayesPGLS (x, plot.type = "traces")}

\arguments{
	\item{ x }{FILL UP}

	\item{plot.type }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

