\name{PrepRegrData}
\alias{PrepRegrData}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{PrepRegrData (data, phylo, formula = NULL, exclSps = NULL)}

\arguments{
	\item{ data }{FILL UP}

	\item{phylo }{FILL UP}

	\item{formula }{FILL UP}

	\item{exclSps }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

