\name{print.BayesPGLS}
\alias{print.BayesPGLS}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{print.BayesPGLS (x)}

\arguments{
	\item{ x }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

