\name{summary.BayesPGLS}
\alias{summary.BayesPGLS}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{summary.BayesPGLS (x, ...)}

\arguments{
	\item{ x }{FILL UP}

	\item{... }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

