\name{summary.potInflObs}
\alias{summary.potInflObs}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{summary.potInflObs (x, digits = 2)}

\arguments{
	\item{ x }{FILL UP}

	\item{digits }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

