\name{RunBayesPGLS}
\alias{RunBayesPGLS}
\alias{RunBayesPGLS.default}

\title{
 FILL UP
}

\description{
 FILL UP
}

\usage{\method{RunBayesPGLS}{default} (formula, data, phylo = NULL, estLambda = TRUE, niter = 30000,  burnin = 10001, thinning = 10, nsim, ncpus, exclSps = NULL)}

\arguments{
	\item{ formula }{FILL UP}

	\item{data }{FILL UP}

	\item{phylo }{FILL UP}

	\item{estLambda }{FILL UP}

	\item{niter }{FILL UP}

	\item{ burnin }{FILL UP}

	\item{thinning }{FILL UP}

	\item{nsim }{FILL UP}

	\item{ncpus }{FILL UP}

	\item{exclSps }{FILL UP}

}

\details{
FILL UP 
}

\value{
	\item{fill up }{FILL UP}
}

\author{Fernando Colchero \email{colchero@imada.sdu.dk}}

\seealso{FILL UP}

\examples{FILL UP}

\keyword{FILL UP}

